pyproject.toml
[tool.poetry]
name = "conversational_ai_library"
version = "0.1.0"
description = ""
authors = ["Radhakrishna Sanka <rkrishnasanka@gmail.com>"]
readme = "README.md"

packages = [
    { include = "nlqs" },
    { include = "expert_system" },
    { include = "tot" },
    { include = "utils" },
    { include = "tog"},
    { include = "state_machine" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
python-dateutil = "^2.8.2"
setuptools = "^65.5.0"
pypdf = "^3.10.0"
pymongo = "^4.3.3"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
chromadb = "^0.4.8"
langchain-community = "^0.3.0"
langchain-core = "^0.3.0"
langchain-openai = "^0.2.0"
pandas = "^2.2.2"
langchain = "^0.3.0"
# TODO: Update the version for pyscopg2-binary in the future right now we are constraining by the version of psycopg2
sqlalchemy = "^2.0.35"
tqdm = "^4.66.5"
# TODO: Workaround for the issue with torch 2.4.1 on macos on intel
torch = "2.0.0"
sentence-transformers = "^3.2.0"
transformers = "^4.45.2"
langchain-huggingface = "^0.1.0"
xformer = "^1.0.1"
openpyxl = "^3.1.5"
psycopg2-binary = "^2.9.9"
langchain-chroma = "^0.1.3"
seaborn = "^0.13.2"
tabulate = "^0.9.0"
pypdf2 = "^3.0.1"
spacy = "^3.8.4"
python-levenshtein = "^0.27.1"

[tool.poetry.group.dev.dependencies]
flask_testing = "0.8.0"
coverage = "^4.0.3"
nose = "^1.3.7"
py = "^1.4.31"
randomize = "^0.13"
tox = "3.20.1"
pluggy = "^1.2.0"
black = "^23.9.1"
mypy = "^1.5.1"
isort = "^5.12.0"
pandas-stubs = "^2.2.2.240603"
pytest = "^8.2.2"
pytest-dotenv = "^0.5.2"
pylint = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pylint]
max-line-length = 120
